Beta Login
+ New login page
+ Beta version everything
Demo is here: http://45.62.98.190/?course=csci1200
- This login thing still need a readme which I will make it up soon.
- There's a problem I haven't fixed, if anyone has some spare time feel
free to fix it. The homework server needs to _GET the course info from
the URL, for exemple "?course=csci1200", but login page is separate from
index page, the last part of url is passed to the login page and needs
to be passed back to the index.
-  Jesse wants it to be visually more like the default theme on the
homework page, if anyone is good at graphic design, go for it