Edit from last submission & Single file deletion
Main changes:
1. The dropzone now displays files submitted last time (before the dash
line) and files chosen for the new submission (under the dash line). By
clicking submit, files above the dash line will be copied from folder of
previous version and files below the dash line will be uploaded.

2. In addition to the DeleteAll button, files can be deselected by
clicking on the file name label.

3. Checks for valid submission added. If nothing has changed in any of
the buckets, or if all the buckets are empty, it is not a valid
submission.

4. Changes made to the buckets only allowed a) when no submission b)
when submission cancelled c) when in the active version and active is
the highest version.

5. Fixed the zip_close position that Matt mentioned.