Multi-parted assignment with drag and drop file upload feature (#374)
* Added history of active version in user_assignment_settings.json

* improved readability

* Removed multiple timezone settings

* Drag and drop file upload feature added.

* Enabled browsing files by clicking on the drop zone. Added checks for submission version and folder upload.

* Implemented multi-part homework upload

Currently a num_parts variable in class.json under each assignment sets
the number of parts. If the variable is not set, then num_parts defaults
to 1. If num_parts > 1, files dropped in drop zone for part n will be
uploaded to folder version_path/partn. If num_parts == 1, files will be
uploaded to folder version_path.

* Edit from last submission & Single file deletion

Main changes:
1. The dropzone now displays files submitted last time (before the dash
line) and files chosen for the new submission (under the dash line). By
clicking submit, files above the dash line will be copied from folder of
previous version and files below the dash line will be uploaded.

2. In addition to the DeleteAll button, files can be deselected by
clicking on the file name label.

3. Checks for valid submission added. If nothing has changed in any of
the buckets, or if all the buckets are empty, it is not a valid
submission.

4. Changes made to the buckets only allowed a) when no submission b)
when submission cancelled c) when in the active version and active is
the highest version.

5. Fixed the zip_close position that Matt mentioned.

* Added file size check for previous files

* Fixed page redirecting & invalid assignment id

* Save files in part folders regardless of num_parts

* Fixed drag and drop events on labels for file names