Changes to be committed: 	modified:   admin-single-user-review.php
KNOWN BUG: Should there be an "upsert" on a regsitered section record for anyone who is not in the students group, the system will (incorrectly) state "User info is updated", but registration info in the DB will (correctly) remain NULL (enforced at upsert call).

TO DO: Disable ("gray out") Assigned Sections checkbox group when user  group selectbox has Instructor or Student (not grader) selected.   For now, user group is validated so that sections can only be  assigned to graders.  Also, students are forced to have no assigned  sections by removing any existing rows in the database's  grading_registration table.  e.g. a user is erroneously added as  a grader with assigned sections, but is later correctly updated to  be a student -- their assigned sections are automatically removed.