Fix Breadcrumbs on TA Grading Pages
This seems to fix the Breadcrumb Issue on TA Grading Pages without reverting the changes made in #1564. If the z-index is set to -1, the object is too low to interact with. Changing the z-index to 0, puts the header (and footer) high enough to interact with, but low enough that it doesn't affect the Grading Tiles