Breadcrumbs Fixes (#1582)
* Fix Breadcrumbs on TA Grading Pages

This seems to fix the Breadcrumb Issue on TA Grading Pages without reverting the changes made in #1564. If the z-index is set to -1, the object is too low to interact with. Changing the z-index to 0, puts the header (and footer) high enough to interact with, but low enough that it doesn't affect the Grading Tiles

* Another Fix For Grading Page

I forgot to change the z-index for the .progress_bar and .grading_toolbar. Also, the z-index for #header and #footer default to 0, so there is no need for the jQuery call.

* Added Admin Breadcrumbs

* Removed Admin Breadcrumb

* Reverted Changes for #1582

* Add to Revert