More obvious : no submission / late submission / cancelled submission (#1707)
* closes #1681

This is a first draft. Color and message can be changed easily.

* making it more organized

* Added banner

I chose many fasinating colors, but none looks good :(. Good thing it's easily changable.

* script missing a space from previous commit

* undo accidental height change

* pull the latest analysis tools source code

* Fixed "Version Conflict" Bug

Switching the order of the if statements fixed the problem.  Instead of checking whether or not TA has graded them first, check if there is anything to grade first.

* adjust colors

* silence dockerfile error message

* silence dockerfile error message