Moved all 'create' tasks to create method
Now, all of the 'create' tasks are in the create method, leaving only the rubric/rotating graders handling to the previously 'modify' method.  This means only the necessary data is saved when the 'Save Changes' button is pressed. and any data that doesn't yet exist in create mode doesn't get saved.