Per-testcase results in GradedGradeable and global GradingQueue (#2379)
* Split GradedGradeable in two

* Updated DB Queries for new split and efficiency

The queries are now setup to handle either missing TA grading or missing auto grading.  If its missing both, then we have no way of knowing anything.  Also, the gradeable data query can query for multiple gradeables at once.

* Simplified GradedComponent and AutogradingVersion

Construction of these are less restrictive and more convenient

* Added Autograding Config model to new gradeable model

This model is a more explicit representation of the autograding config data that is typically used by the gradeable and its consumers instead of an array.

* Documentation fix

* Moved autograding percent calculation

Its now in the version class, so each version can have its percent calculated and the AutoGradedGradeable juts gets the active version percents. This also requires the version to have access to the graded gradeable, so that was readded to its constructor.

* Gradeable lazy loads teams

In addition to knowing if there are any teams, it is often useful to know what those teams are.  Now, these teams are lazy loaded into the gradeable instance (not serialized in toArray())

* Updated phpdoc property type

* Renamed AutogradingVersion to AutoGradedVersion

This is to be more consistent with the AutoGradedGradeable

* Added getter for GradedGradeable in TaGradedGradeable

* Removed unused private setter from GradedComponent

Graded components don't need to know about the graded gradeable

* Updated PercentGraded calculation

WIth some new information about how peer grades are handled.

* Added SubmissionTextBox class and updated config comments

I completed the documentation for properties on the AutogradingConfig class and added  a simple SubmissionTextBox class to encapsulate the data in a Textbox submission.

* Temp Changes

* Completed autograding testcase/version integration

* Fixed typo and removed header

* Fixed gradeable data selector

I forgot some grouping symbols here.

* fixup me

* Added doc

* Fixed merge conflict remenents

* Added important defaults to autograding config

* Initial work on loadQueue

* Fixed checks for graded component/autograding version

* Improved spacing

* Testcase results serialized in version

* Added missing property tags

* Added missing property tags to GradeableAutocheck

Separate commit since this is modifying an in-use class

* Exceptions consistent with other date exception handling

* Misc documentation fix

* Finished queue status methods

* Moved grading queue to core

Instead of each gradeable version calculating this, it gets lazy-loaded once in the core.  This won't improve performance in most cases, but it is organized better  This also means it can be mocked as part of mocking the core.

* Better array handling for no active version

This probably won't happen, but just in case.

* Removed old code

* Updated documentations for enumerations

* Removed AutogradingVersion

Replaced with AutoGradedVersion

* Fixed queue position offset

* Removed old variables

* More appropriate missing file handling

* Clarified comment question

* Removed out of place gradeable methods

* More removed gradeable changes