Changes to be committed: modified:   site/app/controllers/HomePageController.php modified:   site/app/controllers/admin/UsersController.php modified:   site/app/libraries/database/PostgresqlDatabaseQueries.php modified:   site/app/models/User.php
Logged in user is noted in a DB query comment when updating any user's info.
The comment will end up logged by DB as part of the query.  This is because
all DB connections are made by a singular master user, which has the side
effect of completelt anonymizing who was logged into Submitty at any time
a user's data was changed.

This is to help track changes to preferred name.  Could also be later extended
for other user data change tracking, as needed.

q.v. issue #3273

WIP