[Refactor:SubmittyUtils] Cleanup dateutils.write_submitty_date to be more consistent
This updates the write_submitty_date function to be a bit more consistent and with less
potential foot-guns. First, instead of printing an error to console, but still returning
if given an invalid type, this function will now throw an exception, so that if someone
were to somehow get an integer into this function, we should not return an integer back
and have some other crash later in an unrelated system. Scanning through the codebase,
we always use this function with datetimes at the moment at least. Second, remove the
space between the milliseconds and timezone if using milliseconds to match the format of
the non-milliseconds datetime string, as well as just generally where someone would expect
the "-0400" string. Third, the optional parameter to add milliseconds to the datetime
string was named "microseconds", and so it was renamed to "milliseconds" to be consistent
with what it was actually doing.

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>