[Testing:Git] expand functional tests
The current functional tests for Git in Submitty was limited to just
grabbing repo as instructor and then minimally testing that grabbing
a repo with invalid permissions would generally fail, though without
verifying that it failed correctly. This worked well enough for
verifying apache + sbin/authentication.py was working, but did not
provide adaquete coverage for when the PHP code errored, such as in
the result of #5473. The error on the latter testcase was now a 500
error instead of an unauthorized error, but we did not catch this.

To improve things going forward, this commit adds testing to cover:

1. instructor checking out student repository
2. student checkout out student repository
3. student checking out instructor repository fails with specific stderr
   output

This provides more broader coverage of cases to exercise the PHP code,
as well as make sure we do error appropriately.

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>