[Bugfix:Forum] View posts by users fixed on stats page
On the Stats page, on clicking the "Expand" button, the posts were not displayed, instead it would display errors in the console. There was a tiny error in the forum.js file wherein a line was accessing the thread_titles variable as if it were a key-value array, whereas it was a numerically indexed array.