[Bugfix:Forum] Validate Forum Thread IDs (#8014)
* Validates that Thread ID is Valid

In displayThreadList(), checks to ignore threads with null ids.  In getSingleThread(), checks if thread id is valid (not null, an integer, and the ID to an existing thead).  If not, renders a failed JSON response case.

* Included check for if thread id is an int

Checked if thread id is an int, and fixed spelling

* [Bugfix:Forum] Catch Undefined Thread ID

Added a check to catch unset thread ids

* [Bugfix:Forum] Reorganize Checks

Reorganized how checks are arranged.  Changed the use of null to empty.