Imma pray that this works but I have like no faith
Update submitty_ci.yml

just testing if it is just the migrator but this is probably a bad idea

Revert "just testing if it is just the migrator but this is probably a bad idea"

This reverts commit e7fe865e963445fbac89559d539c15ce474c063e.

Seeing if config is even in the right folder

just checking if there is even a path

Having issues with the other tests, suspect because of recent merges. But can always rollback to the "remove git" test.

seeing how much I can remove - lets go bold first

forgot stuff

seeing if we can reinclude the bash

Cannot find submitty_install.sh so just have to hope I don't need that and testing the rest of it

trying this again to see if error with with config.bin

tired