See if this works
maybe I just need to put postgress back on?

Update submitty_ci.yml

just testing if it is just the migrator but this is probably a bad idea

Revert "just testing if it is just the migrator but this is probably a bad idea"

This reverts commit e7fe865e963445fbac89559d539c15ce474c063e.

Seeing if config is even in the right folder

just checking if there is even a path

Having issues with the other tests, suspect because of recent merges. But can always rollback to the "remove git" test.

seeing how much I can remove - lets go bold first

forgot stuff

seeing if we can reinclude the bash

Cannot find submitty_install.sh so just have to hope I don't need that and testing the rest of it

trying this again to see if error with with config.bin

tired

taking a look

Update submitty_ci.yml

Update submitty_ci.yml

Revert "maybe I just need to put postgress back on?"

This reverts commit 985ec7c0b4884755ae892e7a71f5d806a84279d8.

Revert "Update submitty_ci.yml"

This reverts commit 5250cb914a9878d3d9d6f09b4865dc37526e2e9e.