[Feature:TAGrading] Create Breadcrumb for Grading_Beta (#9454)
### What is the new behavior?
When one clicks the Grade button for an assignment on the Gradeable
page, they are taken to the Gradeable's Details page. If one clicks the
Grade/Score button for a particular student, and this Gradeable is an
Electronic/Rubric (Rubric is the new name we are changing to) gradeable,
the TA Grading page will appear. The URL will look something like this:


`SERVER/courses/TERM/COURSE/gradeable/GRADEABLE_NAME/grading/grade?who_id=ANONYMOUS_STUDENT_HASH&sort=id&direction=ASC`

If `/grading/` in the URL is changed to `/grading_beta/`, a new page
will open. This is the location of the in-development refactored version
of the TA Grading page. The page is currently blank, but the header bar,
for the URL example above, will say,

Submitty > COURSE > GRADEABLE_NAME Grading > Grading Interface ID
Ascending Order.

### What should be tested?
On the Details page, each column's header can be clicked, and their
associated <> arrows can be clicked to change the sorting order from
ascending to descending. If you click a student's assignment and then go
the `/grading_beta/` page, the header will reflect the sorting change.

### Previous PRs
This PR is a restart of the work done by @DefinitiveAbove, and it
replaces PRs #7407, #7812, and #7847. There will be a series of PRs as
the rubric grading code is refactored.

---------

Co-authored-by: Mahi Pasarkar <47994802+mahi-pas@users.noreply.github.com>