[Feature:Developer] Support ARM on vmware (#9540)
### What is the current behavior?
Trying to run vagrant using vmware on ARM does not work as it will
always use the x86 image.

### What is the new behavior?
The ARM image gets specified when needed for the vmware provider. This
will not effect any other provider and vmware would still work fine for
anyone not on ARM.

### Other information?
I tested this since I needed this change to use the VM on my machine.