[Bugfix:CourseMaterials] Change the Beginning of Time (#9732)
### What is the current behavior?
There is currently a bug on Submitty regarding course materials set to
the distance past (See issue #9730). While course materials set to the
distant past still do not work, this PR changes the "Beginning of Time"
upload option to the year 1900 to circumvent the issue for now.

### What is the new behavior?
The "Beginning of Time" upload option works!

This PR closes issue #9733, a checkmark of issue #9730.