[Bugfix:TAGrading] Checkpoint Grading Empty Section (#9804)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
If you try to load the checkpoint grading page with an empty section, a
frog robot error will occur. This is due to the null coalesce being null
rather than an empty array.

### What is the new behavior?
The page loads normally as expected. The null coalesce is now just an
empty array as it should be.

### Other information?
Tested locally and it does fix the issue.