[Documentation:Developer] Wrap placeholder text in HTML comments (#9899)
### What is the current behavior?
Our issue templates currently include placeholder text. If the
placeholder text is not removed manually prior to opening an issue, the
placeholder text will remain visible once the issue is opened.

### What is the new behavior?
To reduce the amount of spam in our new issues, I have wrapped the
placeholder text in HTML comments. The placeholder text will now be
hidden if it is not removed prior to opening an issue.