[Bugfix:System] Fix session cookie not updating (#9925)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
When the session gets closer to expiration it will update on the
database but not the cookie. Thus the cookie will still expire and log
the user out after the session expiration time. The issue was introduced
in https://github.com/Submitty/Submitty/pull/8620.

### What is the new behavior?
The cookie will now update when the session updates. The user will only
get logged out after 2 weeks of inactivity.