[Refactor:Forum] Replace depreciated DOMNodeInserted (#9874)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
Closes #9868 (DOMNodeInserted depreciation from forum). The forum page
previously used the event DOMNodeInserted to call a function every time
an image was uploaded while creating a post. This gave console warnings
about the event being depreciated.

### What is the new behavior?
The forum now uses a mutation observer to observe an image upload. The
console doesn't issue any more warnings.

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->

When creating a post, adding images, deleting images, and sending one or
more images work as previous, but without the warning of a depreciated
event.