Update "Late Days" page to show late days used for bulk upload gradeables #10051.js
In the calculateLateDays function, I added a check to see if the selected gradeable is a bulk upload and whether late days have been used. If it's a bulk upload with zero late days used, the late days value won't be updated.

In the updateLateDays function, I included information about bulk uploads by appending a is_bulk_upload field to the form data.

Additionally, I added an event listener to the gradeable selection to recalculate late days when the user changes the selected gradeable.