[Bugfix:Forum] Add more empty list checks (#10055)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
There's an issue in the forum that's been affecting FOCS that occurred
when the professor went to the forum and filtered by unread. It happens
when SQL tries to query an empty list of parameters, and I was able to
reproduce this by filtering by unread when I had no unread threads.

### What is the new behavior?
I've added checks for empty lists before all the queries used to display
the thread list. In the event of an empty thread list, the server will
query no data and then iterate over the empty list and display nothing.
I'm able to filter by unread even with no unread threads now.

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->