[Refactor:Developer] Lint remaining JS files (#10102)
### What is the current behavior?
Eslint is currently run on most of our files, but the target JS version
is too low to be able to effectively parse a handful of remaining files.

### What is the new behavior?
This PR changes our JS target version to `es2020`, and finishes linting
the rest of the JS files. All of our JS files are now linted on every CI
run.