[Refactor:System] Switch PDO to DBAL (#10245)
### Please check if the PR fulfills these requirements:

* [x] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
We currently wrap our AbstractDatabase around PDO for both Postgres and
SQLite. We should also be upgrading DBAL to 3 (and eventually 4 since
that was just released) which has been an open PR for a while. Doctrine
ORM requires DBAL because that is what it uses as it's database layer.
We pass in our existing PDO instance to ORM/DBAL to avoid creating a new
database connection.

We are using DBAL 2.13.9.

### What is the new behavior?
This PR changes AbstractDatabase to now wrap DBAL instead of PDO. DBAL
wraps PDO so we essentially just have an extra layer in between
AbstractDatabase and PDO. If we were to keep AbstractDatabase using PDO,
we would need to double our database connections to support DBAL 3 and
higher. This is because DBAL removed the ability to pass in a user
provided PDO instance. So we either would need to double the database
connections or switch our connection. The idea is to maintain just 1
DBAL connection for each the master and course DBs and to just pass that
into ORM now.

We are now using DBAL 3.8.2.

### Other information?
As little has been changed to replace PDO to DBAL in AbstractDatabase.

---------

Co-authored-by: Matthew Peveler <matt.peveler@gmail.com>