[Bugfix:Developer] Initialize "os" key in Docker UI (#10314)
### What is the current behavior?
https://github.com/Submitty/Submitty/pull/10271 missed the "os" key,
which is believed to be the cause of
[this](https://github.com/Submitty/Submitty/actions/runs/8515492568/job/23323056920)
CI failure.

### What is the new behavior?
The "os" key is initialized to null.

The original source of the Docker UI test flakiness is PR
https://github.com/Submitty/Submitty/pull/7932. Notably, the last
comment on that PR discusses the flakiness of the tests, and the
subsequent
[commit](https://github.com/Submitty/Submitty/pull/7932/commits/5acb72fb030ac02cbd54195e786a62dbea96f386)
comments out a test in an attempt to fix the flakiness. In the future,
it would be worthwhile to investigate the commented-out test to see if
it is now reliable.