[Feature:TAGrading] User assignment settings (#8632)
### What is the old behaviour?

The
/var/local/submitty/courses/SEMESTER/COURSE/submissions/GRADEABLEID/STUDENT/user_assignment_settings.json
file contains information about the history of upload timestamps and
changes to the active version (cancel, rollback, etc). This information
can be useful in determining how to grade a student. This file is
currently only accessible if we ssh to the machine and look at the files
directly.

### What is the new behavior?

The user_assignment_settings.json file is now visible in the TA grading
interface and If we are in anonymous mode the file is not available
since it contains the student username.

Closes #8459

---------

Co-authored-by: saumyaborwankar <46644513+saumyaborwankar@users.noreply.github.com>
Co-authored-by: musaabimran <musaabimran2001@gmail.com>
Co-authored-by: Chris Reed <55092742+cjreed121@users.noreply.github.com>
Co-authored-by: Musaab Imran <96174078+musaabimran@users.noreply.github.com>