[Bugfix:Autograding] Copy all named executables to runner (#10173)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
Fixes #10121 . Currently, the autograder only copies .out and .class
files from the compiler to the runner. This means if the instructor
creates a compilation testcase that compiles into anything other than
.out or .class, the file will not be copied to the runner and other
testcases will fail.

### What is the new behavior?
Now, in addition to copying all .out and .class files to the runner, any
file that is listed in the "executable_name" field in the gradeable's
config.json will also be copied to the runner.

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->

This needs more testing. I've been able to verify that it works when a
single executable is produced, but I haven't been able to test whether
it correctly handles an array of executable names. I'm currently working
on a java example for this, and I'll upload it as a separate PR when
it's done.