[Feature:Submission] Student Git Repo Creation (#10164)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
If a student is added to a course through the autofeed script and a VCS
gradeable is created before they entered the course, that user will not
have a VCS repo created for them. They will be prompted that they can
clone on the submission page but any attempts to clone will fail.

### What is the new behavior?
The website will now check if the repo exists. If it doesn't then it
will prompt the student to create the repo. Upon clicking the button, a
job will be added to the daemon queue and will go create any missing
repositories.

### Other information?
A new group was added for supporting this functionality.