[Bugfix:System] Fix VCS permission on install (#10404)
### What is the current behavior?
PR https://github.com/Submitty/Submitty/pull/10164 added a migration
which changed the permissions of the vcs directory and subdirectories.
In one of the install scripts it changes the permission back to the old
value which breaks the website submission. Also it only uses the student
ID for checking the repo path.

### What is the new behavior?
The install script was edited to set permissions correctly. Now it will
use student ID or team ID for the repo path to support team gradeables
too.

### Other information?
Tested locally and submission pages works correctly after install