[Bugfix:System] Adduser group optional (#10479)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
In https://github.com/Submitty/Submitty/pull/10133, a new group argument
was added to the adduser script. However, it is not an optional argument
since it is missing `--`. The create course script will error as is and
not add the instructor to the course.

### What is the new behavior?
`--` has been added in front of the argument so this is now an optional
argument. The create course script does not fail anymore.

### Other information?
Tested locally and it fixed it.