[Bugfix:Forum] Upduck counter wrap to next line bug (#10392)
### What is the current behavior?
When a user edited a post in forum or user's display name is too long
(for smaller screen sizes), Upduck counter will wrap to the next line.
But, other elements (even the Upduck button doesn't.) don't go to next
line.


![image](https://github.com/Submitty/Submitty/assets/68075099/3260be2c-139d-40d3-9ba4-27f8e54d163b)

### What is the new behavior?
Nested the Upduck button and counter inside a span and again nested it
inside span.post-action-container. Now if the width exceed the space,
whole span.post-action-container will go to the next line. Borders are
optional. Added it to improve the UI.

#### Dark mode

![image](https://github.com/Submitty/Submitty/assets/68075099/faddbb39-34d3-4237-81c8-621565973750)
#### Light mode

![image](https://github.com/Submitty/Submitty/assets/68075099/bc9a3259-d2c2-48c3-8bfe-6a642ab1ae5e)

closes #10367 

### Other information?
I think there is no potential issue of nesting Upduck button and counter
inside a span. Button and counter are working correctly.

---------

Co-authored-by: Michael Papadopoulos <micpap25@gmail.com>