[Bugfix:System] Preserve submitty config keys (#10401)
### Please check if the PR fulfills these requirements:


### What is the current behavior?
Currently, when the CONFIGURE_SUBMITTY.py script is run, the
system_message field located in /usr/local/submitty/config/submitty.json
is not preserved. This can be annoying for sysadmins who don't want to
have to rewrite their system_message everytime after they run
CONFIGURE_SUBMITTY.py.
Closes #8151 
### What is the new behavior?
If the submitty.json file has defined a system_message, field, it is
saved into the config of the new file.

### Other information?
You can test this by running the CONFIGURE_SUBMITTY.py script located in
/usr/local/submitty/GIT_CHECKOUT/Submitty/.setup/CONFIGURE_SUBMITTY.py

Please try all sorts of edge cases. 
Examples:
Delete the submitty.json file and run
Change the field and run and make sure that it is preserved
leave the field empty
NOTE: we are assuming that the file is valid json, that is the user's
responsibility

---------

Co-authored-by: Jaeseok Kang <ziesski@gmail.com>
Co-authored-by: Sátvik Karanam <89281036+skara9@users.noreply.github.com>