[Bugfix:Developer] Fix stale PR CI retagging after 2 days (#10534)
### What is the current behavior?
If a PR is abandoned, then picked up by a new owner, and then left
inactive for 2 days, then the PR is re-tagged as abandoned. This is
because the warning has already been issued the first time and the
script checks if it exists at all before tagging as abandoned.

### What is the new behavior?
This PR makes the script only check the latest comment for the
inactivity warning before tagging so that this doesn't happen.