[Feature:Calendar] Month and year dropdown menu (#10249)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [x] Documentation has been updated/added if relevant
* [x] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
Closes #10242, closes #10246. Currently, users have to repeatedly click
the arrow buttons to move the calendar by a month. When going back more
than a few months, this is tedious.

### What is the new behavior?
This PR implements a dropdown menu (4 years in the past, 1 year in the
future) so that users can more easily go back and forwards in time on
the calendar. It uses cookies to keep the calendar on the same state
when reloading the page as well.


![image](https://github.com/Submitty/Submitty/assets/122962727/cfd26f36-d0b7-4092-b0b1-a7378906a0fa)

### Testing
This was tested on the VM as instructor with many different months and
years selected.

---------

Co-authored-by: Zheyu Deng <122962727+ZheyuDeng@users.noreply.github.com>
Co-authored-by: Zheyu Deng <sony1119266216@gmail.com>