[Feature:Developer] run ntp on submitty_install on vagrant (#10587)
### What is the current behavior?
ntp does not reliably keep the time synced on all vagrant VMs
(problem seems to be worse on ARM macs?)
closes #9584 

### What is the new behavior?
This PR expand the force running of ntp in submitty_install from only on
UTM VMs to all vagrant VMs.

related to #10458 and #10469

This problem is a long-standing, known issue with VMs. Previously it was
solved using Guest Additions (which have other long-standing, known
complications keeping versions up-to-date), but now that we have finally
removed the need for Guest Additions in our Submitty VMs we don't want
to re-add those headaches just to fix the occasional clock skew problem.
This fix is imperfect, the skew is only corrected when install_submitty
is run. An alternate (but still imperfect) solution would be to add a
cron job to force clock updates. I think this PR is the right
compromise.

Co-authored-by: Barb Cutler <Barb Cutler>