[UI/UX:TAGrading] Disable notebook fields when grading (#10604)
### What is the current behavior?
The answer fields in the notebook panel of the TA grading interface are
only disabled for submissions other than the active submission. These
fields should always be disabled in the TA grading interface.

### What is the new behavior?
All notebook fields are disabled in the TA grading interface.  In

### Other information?
In the future, it would be good to put together a set of comprehensive
tests for the notebook panel of the TA grading interface.