[Bugfix:Developer] Fix CI Flag and Ignore Flag File (#10672)
### Please check if the PR fulfills these requirements:

### What is the current behavior?
The CI flag works on CI, but not locally because of a formatting issue. 

### What is the new behavior?
The CI flag is now fixed, and accurately skipping the NTPD setup on CI,
but allowing it to run on local machines.
Also, the flag file was added, so all users who didn't delete the file
were running it as if they were in the CI, so this file has been added
to the .gitignore file, so if someone is testing the feature, it won't
be committed and sent to all users who pull.