[Feature:Submission] Hide Accessibility Text for Disabled (#10649)
Fix #10643

### Please check if the PR fulfills these requirements:

* [X] Tests for the changes have been added/updated (if possible)
* [X] Documentation has been updated/added if relevant
* [X] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
CodeMirror currently displays accessibility help text above all code
boxes. This text is confusing in places where the code box is disabled
and it is impossible to navigate to it.


### What is the new behavior?
This commit updates the `getLargeCodeMirror` function to conditionally
display the accessibility help text only when the CodeMirror instance is
enabled. This prevents confusion in scenarios where the code box is
disabled and navigation is not possible.


![image](https://github.com/Submitty/Submitty/assets/34382211/dbf21029-63d0-4075-8bea-1d0391a4749b)

---------

Signed-off-by: Sahil Suman <sahilsuman933@gmail.com>