[Feature:SubminiPolls] Custom answers for polls (#10197)
### Please check if the PR fulfills these requirements:

* [x] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [x] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
Currently, students cannot post custom responses to a given poll on
Submitty
 
### What is the new behavior?
Fixes  #6236

Instructors may now allow students to submit custom options that other
students can also select for any poll question type. Instructors may
also allow these custom options that students submit to count as credit
as they would normally for a poll option. Because a large number of
custom options can be submitted, a new checkbox was added in the poll
form to be able to toggle all options as correct or not for ease of use.

### Other information?
To test this feature, first run migrations for the course database, and
then when creating a poll, toggle the checkbox stating "Allow poll to
accept custom answers from students" to allow students to enter their
custom responses. Log in as a student to post your custom response to
the poll question. Any other student account may view the option and
choose it if they wish. These custom options may only be deleted by the
student if they are the only person who selected them or by instructors
if no students have selected the custom option. In the extreme case of
an inappropriate option, instructors can still edit the option from the
standard edit poll form.


![poll-custom-option](https://github.com/Submitty/Submitty/assets/94331560/8a7f6512-4496-433a-9464-62f27bfcc6ec)


![poll-custom-submit](https://github.com/Submitty/Submitty/assets/94331560/452555a3-4dad-498c-9f2d-008563b8f386)


![poll-custom-instructor](https://github.com/Submitty/Submitty/assets/94331560/b4cf5363-c16c-4599-a82c-e24098d20346)


### Other information?
When this PR is merged, the following issues should be added:

- hide delete button for responses that have at least one student
selecting it on both the student and the instructor view
- show id of student who added response to instructor view
- limit total number of responses

---------

Co-authored-by: zacharymnp <56311867+zacharymnp@users.noreply.github.com>
Co-authored-by: powe97 <116031952+powe97@users.noreply.github.com>
Co-authored-by: William Allen <16820599+williamjallen@users.noreply.github.com>
Co-authored-by: Jaeseok Kang <123261952+ziesski@users.noreply.github.com>