[Refactor:Developer] Mv tutorial course yml to Tutorial repo (#10761)
### What is the current behavior?
See Submitty/Tutorial#45. Currently the course config yml for each
preset course in the VM is in the Submitty/Submitty repo.

### What is the new behavior?
This PR moves the `tutorial.yml` course config to `course_config.yml` in
the Tutorial repo. This will be part of a refactor to move
repo-dependent course yml files into their respective repos.