[Bugfix:Developer] Fix draft arranger for old PRs (#10798)
### What is the current behavior?
The auto draft mover fails.

### What is the new behavior?
Fixes one more bug introduced in #10775. By default, the github CLI only
returns the most recent 30 PRs in queries, so this action would fail for
any PR older than that due to not being able to find the PR ID. This PR
changes the query to search the most recent 9999 PRs instead, which
should be sufficient.