[Feature:Developer] Move review requested PR to In Review (#10848)
### What is the current behavior?
Currently, there is no project board automation for moving PRs to the
"In Review" column. It is very easy to forget to do this, so some PRs
get lost as a result.

### What is the new behavior?
This PR automates moving a PR to "In Review" when a reviewer is
requested, for PRs in the Submitty/Submitty repo.

After this PR, this is the project management flowchart. This PR adds
the "Reviewer assigned" automation.


![current](https://github.com/user-attachments/assets/ea3d4e73-c25a-4991-8eaa-bdb081503062)

---------

Co-authored-by: William Allen <16820599+williamjallen@users.noreply.github.com>