[Refactor:SubminiPolls] Refactor Frontend PollTimer Cypress test (#10957)
closes #10927  

### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?

Currently, the cypress test is testing the front end displaying "Poll
Ended" more as status than understanding if the Poll Ends after a given
time interval. The cypress test isn't accurately testing the 'time'
functionality of the front end.

### What is the new behavior?

An additional front-end check was added after the wait statement to
ensure that the front end was synchronized with the back end after the
given interval.