[Bugfix:Developer] Give Cypress Artifacts Unique Names (#11000)
### What is the current behavior?
When a Cypress CI test fails, it tries to upload artifacts named
`cypress-videos` and `cypress-screenshots`, however if multiple of the
cypress matrices fail, then only the first artifact gets uploaded, as
they all share the same name.


### What is the new behavior?
The Cypress artifacts are now named based on the 'container' or
'folder', like `Cypress-TAGrading-Screenshots` or
`Cypress-Feature-Videos`. This will allow for developers to see all of
the tests that failed in CI, not just the first one that completed.