Added click component to received marks list and set default showing to above 0
In this change, I ensured that the received-marks-list container only displayed the changes in the previous PR if there was a graded component already. If there was a graded component, then it shows all marks whether checked/unchecked.

Also, I made it so that when the container is clicked, it will open the entire container, fixing confusion experienced by TAs.