Improving Cypress Test Logic
Currently, the cypress tests check if each component contains text before a graded component is present, which does not make sense since the component is empty in this case per the previous commit. So, it has now been made to check if every mark shows in closed view i.e after the save button is clicked.