[Feature:TAGrading] Resolve Marks Confusion in Closed View (#11032)
Closes #11028 

### What is the current behavior?

Currently, TAs are experiencing some confusion with the feature
implemented in PR #11010. The confusion mainly lies with clicking
components when the entire marks lists show and interface problems.

### What is the new behavior?

This PR introduces two separate changes. For one, when a component has
no graded marks, it does not show the received-marks-list container at
all (as shown below):


![image](https://github.com/user-attachments/assets/f4872987-ab72-4d16-827e-3f6cc9a351ae)

Secondly, when a graded component is clicked, it will always open the
container regardless of which area the TA/grader clicks. It could be the
center, on the check marks, the right/left hand corners, etc.

### Other information?

To test, create or go to an already existing gradeable and open the
grading page. Then, verify that the above changes work as intended.