[Bugfix:SubminiPolls] Poll Timer Cypress Student View (#11123)
Fixes #11037 
### What is the current behavior?
If the Poll timer didn't work in the student view, the Cypress test
wouldn't have found out if there was an accessibility error in the
student access view poll but would have caught the mistake if it messed
up the instructor access page.

### What is the new behavior?
There are valid checks for both student and instructor to ensure it
catches any issues with the poll timer.