[Refactor:Forum] Use Doctrine to get forum post history. (#10983)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->

Doctrine is a database library including an ORM and DBAL. We use it in
several areas, such as polls, but have not yet added it to the forum.
This hurts the readability and design of the forum code.

### What is the new behavior?
No functionality should change that is visible to the user. Expands on
existing Doctrine infrastructure for forum and uses it to get post
history. This is the first PR of several to come that will convert the
entire forum to Doctrine.

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->

PHPStan complains (and is silenced) about the use of generic Doctrine
types in several sections of the code. I don't believe we will/should
attempt to resolve those so rather than adding these errors to the
baseline as they accumulate through increased Doctrine use, so I want to
not check them.