[Bugfix:Forum] Fix error on duplicate bookmark request (#11154)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
If a user sends multiple requests to bookmark a specific thread (before
the page reloads), the requests after the first will cause a keyerror in
the database and throw an exception.

### What is the new behavior?
Ignore these duplicate requests; if the row to be inserted is in the
database already, do nothing.

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->

You can reproduce the error by commenting out the
`window.location.replace` line at forum.js:1883 (function
bookmarkThread). Clicking the bookmark button multiple times should
cause subsequent clicks to return a frog robot on main, but they should
succeed with this change.