[Bugfix:Developer] Fix Vagrant Box Packaging (#11150)
### What is the current behavior?
Vagrant has been migrated to be completely a part of Hashicorp, and now
uses a different token to allow publishing to Hashicorp Cloud, instead
of Vagrant Cloud. We did not have this token as all we used to need was
a Vagrant Cloud token.

### What is the new behavior?
A Hashicorp Cloud token has been added and retrieved, which allows the
vagrant box to be uploaded.

Eventually we will refactor the workflow file to use fully the Hashicorp
API or CLI, but as of now the instructions are not clear, and it is
still early after release.