[Refactor:Autograding] Config-form <script> blocks now in js (#11159)
### Please check if the PR fulfills these requirements:

* [ X] Tests for the changes have been added/updated (if possible)

### What is the current behavior?
Currently, there are jquery <script> blocks in the UploadConfigForm.twig
file. This isn't considered good practice. This PR is an extension of
previous PRs I've made, which change the autograder page to immediately
upload a chosen config file rather than waiting for the user to do it
manually.

### What is the new behavior?
Aside from chosen config files being immediately uploaded, the behavior
of the rest of the page should be the same. I've taken the <script>
blocks and put them inside a js file, which is loaded in via
"addInternalJs()" within the GradeableView.php file.

### Other information?
For reviewers, check the make sure that chosen file are immediately
uploaded. Also check that editing/deleting uploaded config files still
functions as intended, so as to make sure I didn't break anything when I
refactored them into a js file.

Extension of PR #11139. Fixes issues #10847 , #11143 , and #11152

---------

Co-authored-by: ION606 <orenni@rpi.edu>
Co-authored-by: ION606 <ion606@protonmail.com>