[Refactor:TAGrading] Removing .html() from server.js (#10980)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->

Using `.html()` or `innerHTML` can introduce security vulnerabilities,
especially related to Cross-Site Scripting (XSS) attacks, as they allow
arbitrary HTML to be injected, potentially including malicious scripts.

### What is the new behavior?

It's safer to use `.text()` in jQuery or `.textContent` in plain
JavaScript when setting or updating text content.

### Other information?
We will be covering our security standards in few of next week short
presentation. This is one of the common mistakes that devs make. This PR
is to replace some of `.html()` with `.text()` and reduce developers
using outdated code as reference.

---------

Co-authored-by: williamschen23 <chenwill2005@gmail.com>