[Bugfix:SubminiPolls] Make poll histogram numbers categorical (#11173)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [x] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
When a Submini poll contains a string of numbers (e.g. `1 2 3 4 5 6 7
8`), Plotly ignores whitespace and incorrectly treats the option as a
large number. This is because Plotly is not interpreting the data as
categorical, and is creating a regular bar chart instead of a histogram.

### What is the new behavior?
This PR tells Plotly to treat all Submini poll options as categorical
data, so there is no number parsing.

Resolves part of #11115 

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->
**Before:**

![image](https://github.com/user-attachments/assets/837e1775-bd02-44b2-9b07-1ed17ae3c030)
**After:**

![image](https://github.com/user-attachments/assets/f9d0127d-66f0-432f-b752-39bc836d3ff5)

**Control Before:**

![image](https://github.com/user-attachments/assets/e54cad67-23f7-4df0-8f08-8df306a8878b)
**Control After:**

![image](https://github.com/user-attachments/assets/999387d1-e0b1-4f47-9e83-58eb17256661)


This does NOT fix the issue where double quotes are HTML encoded, due to
unspecified behavior on how surrounding quotes should be handled. I
think that should be a separate GitHub issue. (see:
https://github.com/Submitty/Submitty/issues/11115#issuecomment-2452844438)

Automated test cases may be needed?

---------

Co-authored-by: zacharymnp <56311867+zacharymnp@users.noreply.github.com>