[Feature:InstructorUI] Add ability to delete docker imgs in UI (#8845)
### What is the current behavior?
Fixes #8524. Currently only adding a Docker container is possible
through the GUI. If the instructor needs to delete an image, it is
possible only through the command line.

### What is the new behavior?
Users can now use the "Remove" button to delete an unnecessary Docker
container. The Docker UI Cypress test has been expanded to test this
functionality.


![image](https://github.com/user-attachments/assets/c8020393-2e54-4070-b3ab-23567d7af36c)

---------

Co-authored-by: Barb Cutler <bmcutler@users.noreply.github.com>
Co-authored-by: Cameron Peterson <46759635+IDzyre@users.noreply.github.com>
Co-authored-by: DarthNyan <tatewhiteberg@gmail.com>
Co-authored-by: Tate Whiteberg <89590361+DarthNyan@users.noreply.github.com>
Co-authored-by: William Powe <116031952+powe97@users.noreply.github.com>
Co-authored-by: Jaeseok Kang <123261952+ziesski@users.noreply.github.com>
Co-authored-by: Jaeseok Kang <ziesski@gmail.com>