[Bugfix:Forum] Dark mode working for category filters (#11181)
The original pull request changed the background of the buttons to gray
when in dark mode, but did not change the text color. This decreased the
contrast, making it hard to read the text. As a solution, I forced the
dark mode text to be white with !important. I understand the use of
important is discouraged, so I will clean up the file which already
includes 20 of them in a future PR.

**Current (Main) Light Mode (remains unchanged):**

![current_lightmode](https://github.com/user-attachments/assets/b4417b6d-91f2-4dfe-9c4f-9b485efbb003)

**Current (Main) Dark Mode:**

![current_darkmode](https://github.com/user-attachments/assets/c07f6bfc-9215-4414-9342-ac3f5b5c1bdc)

**Original Pull Request Dark Mode:**

![old_darkmode](https://github.com/user-attachments/assets/d756704e-283e-49fe-bffd-c314cebdb1b5)

**New Dark Mode:**

![new_darkmode](https://github.com/user-attachments/assets/1e68df8a-bf02-4024-aefa-858e22d72ee4)

---------

Co-authored-by: Justin Manion <jmanion32@gmail.com>