[Refactor:Developer] PHPStan: ignore PHPDoc types (#11336)
### What is the current behavior?
PHPStan trusts PHPDoc comments by default. While this can be good if
done carefully, the PHPDoc types in the Submitty codebase are not
sufficiently thorough to take advantage of this. Instead, our incorrect
annotations can cause PHPStan to come to incorrect conclusions about the
type constraints present.

### What is the new behavior?
This PR disables PHPStan's `treatPhpDocTypesAsCertain` rule, causing it
to rely solely on types which are guaranteed to be correct based on the
language itself.