[Bugfix:Forum] Fixed moving textbox. (#11330)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [x] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
Let's say users A and B are discussing over a thread, and while user A
is typing a message, user B drops two messages. Due to this, the textbox
of user A moves down.

### What is the new behavior?
Fixes #10934 
The textbox for user A doesn't move if a new message appears in the
thread. In addition to that, the page automatically scrolls down to the
bottom of the thread once a new message arrives.


https://github.com/user-attachments/assets/34335f97-66ce-4cae-8e7c-1d3bd8d2496a


### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->