[Bugfix:Forum] Forum Network and Reply Level Issues (#11365)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [X] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?

When a user is on the same thread as another connected to the same
WebSocket, after one of the users sends various new posts to the thread,
once the other tries to click the thread on the left-hand thread list,
an error is alerted with no further updates.

I also realized the incorrect `reply_level` for new posts is being sent
over the WebSocket, leading to an invalid tree structure and/or
rendering errors for further modification of posts. The following
displays the right-hand user receiving the invalid ordering from the
WebSocket.


![image](https://github.com/user-attachments/assets/b0be33bf-2bb6-463f-8ee4-94779850451c)


### What is the new behavior?

No alert is shown for the potential network issue caused by the
WebSocket updates interfering with the AJAX request cycle. Ordering
posts from the WebSocket matches the expected output on both sides.


![image](https://github.com/user-attachments/assets/c4c3c968-4b62-4122-8ee9-786537b9b13d)


### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->

Fixes #10933