[Bugfix:Forum] Fix Upduck Count Stat (#11442)
### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
The current query to get all upduck is wrong. This could be simply
observed by logging in as instructor and upducking a post not by
instructor. This would cause the upduck count not go up

http://localhost:1511/courses/s25/sample/forum/stats

This is 1/2 of fixing the forums, the other one is below:
#11441 

### What is the new behavior?
Fixes the query and the related cypress tests

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->
This is the first part of many changes to the forums and upduck tests.