[Refactor:Forum] Mark post unread with Doctrine (#11388)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
Doctrine is not implemented for the markPostUnread method in
ForumController.

### What is the new behavior?
The markPostUnread method is optimized to use Doctrine.

### Other information?
The mark post unread option under the option dropdown of any post in a
thread should be functional. After marking a post as unread and moving
to a different thread and back, the UI should highlight the post as
unread.