[Refactor:Forum] Mark post unread with Doctrine (#11388)
* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

Doctrine is not implemented for the markPostUnread method in
ForumController.

The markPostUnread method is optimized to use Doctrine.

The mark post unread option under the option dropdown of any post in a
thread should be functional. After marking a post as unread and moving
to a different thread and back, the UI should highlight the post as
unread.