[Refactor:Forum] Remove !important from forum.css (#11349)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
Closes #11347 
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
forum.css currently contains many instances of !important which forces
elements to follow specific styles regardless of other CSS rules.


![image](https://github.com/user-attachments/assets/51394383-fd88-4607-a2e2-214746159ca4)


### What is the new behavior?
Removed all instances of !important from it to follow the Standard
Specificity Rules which makes it easier to debug and maintain the code.


![image](https://github.com/user-attachments/assets/996bed29-7e9c-494e-bd36-026ec277444a)


### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->
Added the screenshots to verify that the Forum looks exactly same before
and after the changes are made.

---------

Co-authored-by: Justin Manion <jmanion32@gmail.com>