[Bugfix:Forum] Separate and fix inconsistent forum test (#11441)
### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
Mentioned in #11407 by @IDzyre 
This is 2/2 of fixing the forums test. The following needs to be merged
first:

- [x] #11442

### What is the new behavior?
Moves the upduck to the new file, and in addition, uses the current
threads to test upducks. Refactor it to not use wait statements.
Will fix the inconsistency of the forum thread merge and deletion.

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->