[Bugfix:InstructorUI] Notify on Self Registration (#11451)
### What is the current behavior?
Currently all instructors in Submitty are receiving emails when a user
self registers for a course. This is because the function used is just
getting all of a certain type of user.
Closes #11434 
### What is the new behavior?
This function has two more parameters added, allowing for the ability to
specify if you want that group for a specific course and term.

### Other information?
This changes the SelfRejoin function as well, as that is supposed to
notify the instructors of that given course, not all instructors of
Submitty.
The original use of this function remains unchanged, as `term` and
`course` are optional parameters, and the query defaults to being the
same query as before.

---------

Co-authored-by: Barb Cutler <bmcutler@users.noreply.github.com>