[Bugfix:InstructorUI] Fix When Default Section is Null (#11486)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
The sample course 'testing' currently is supposed to allow for self
registration, however in doing so it does not set the default section
id, and breaks the functionality somewhat, as users who register are put
in the null section.
Related to https://github.com/Submitty/Submitty/issues/11471. This does
not fix the issue as described, users who are set to the null section
will not be able to register, but it gives the users a warning message,
and displays a different default section to instructors.

### What is the new behavior?
The sample course scripts now accurately set the default section, and if
an instructor uses the create_course.sh script, and does not set a
default section, it will now display a different message to the
students, and not allow them to register. Also, the place to set the
default section now shows 'NULL' if the section is null, instead of
showing the first given section.