[Bugfix:Forum] Fix the Thread Lock Date error (#11531)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
By default, discussion forum threads have a NULL thread lock date, and
any user can post to the thread.
If the instructor specifies a thread lock date, students should not be
able to post to the thread after that date.

However, it seems that if the thread lock date is non-NULL, students are
blocked from posting to the thread. Even if the date is far in the
future.

Also, it's not possible to edit a thread and remove a thread lock date
-- e.g., set it back to NULL. You can change the thread lock date to
another date.
### What is the new behavior?
Thread Lock Date Validation: If the thread lock date is set to a valid
future date, students should are able to post before that date
Thread Lock Date Removal: Now can edit a thread and set the thread lock
date back to NULL should be restored.

### Other information?
Fixes #11499