[Refactor:System] Remove checks for empty preferred names (#11376)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
Preferred names used to be allowed to have a value of NULL or `""` which
would both indicate that the user does not have a preferred name and
wishes for their legal name to be used instead. Once #11359 gets merged
that will no longer be the case and only NULL values will indicate that
the user does not have a preferred name.

### What is the new behavior?
This PR removes all of the existing checks for `""` as a preferred name
because those checks are no longer possible.

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->