[UI/UX:InstructorUI] Dark Mode for Customizing Categories (#11541)
Currently, there is no dark mode implementation for customizing course
categories in the discussion forum, causing it to stick out and making
the delete icon invisible:

![main-category-lm](https://github.com/user-attachments/assets/0908902e-73ea-4472-9210-d920b517da96)

![main-category-dm](https://github.com/user-attachments/assets/fd2a77c6-776f-413f-99ce-f8ee9362f1e3)

Implemented dark mode for customizing course categories:

![pr-category-lm](https://github.com/user-attachments/assets/c3a144e7-a268-4322-af00-62cbcc8ff97b)

![pr-category-dm](https://github.com/user-attachments/assets/fbae5420-981b-4e15-8535-c27da9805424)

Notice that the text color no longer matches the category color. This is
intentional as the contrast against the background would only be good in
light mode.