[Testing:System] Linter Test for !important (#11494)
### What is the current behavior?
Currently, there is no check for !important usage in CSS. It is best
practice to avoid the use of this !important because it disrupts CSS's
order of specificity, and makes maintenance more difficult in the
future.

### What is the new behavior?
Added a stylelint rule limiting !important occurrences in .css files.
Added inline ignores to all existing occurences of !important that would
otherwise be flagged by the rule.

---------

Co-authored-by: Troy-bailan-King <97757033+Troy-bailan-King@users.noreply.github.com>