[Feature:SubminiPolls] Add WebSocket to Polls (#10253)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [X] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
Currently, the histogram of the poll page on the instructor side does
not update dynamically. Instructors must refresh the page to get the
most up-to-date poll information to be reflected on the histogram. Also,
any updates on the current open poll will not be reflected on the
student interface unless stated directly by the instructor.

### What is the new behavior?
Instructors can now see updates reflected on the histogram as students
submit their poll responses. On student submission, a WebSocket message
will be sent only containing updates for the most recently changed
option response counts. Students will also receive WebSocket messages
when a poll has closed, re-opened, or has made significant changes.
 
### Other information?
This is a small change to the polls feature of Submitty. I tested by
submitting various changes in polls side by side as instructor and
student. As a student submits a change in a given poll, the histogram
would update on the instructor side in constant time, reflecting current
response counts. I also made various changes to the poll while a student
was connected as shown in the short demonstration video below. This
feature works toward completing
[#5913](https://github.com/Submitty/Submitty/issues/5913).



https://github.com/Submitty/Submitty/assets/94331560/b154b6e7-592d-4301-8079-4404283acbdc

---------

Co-authored-by: powe97 <116031952+powe97@users.noreply.github.com>